Warning fixes 64-bits fixes
git-svn-id: svn+ssh://svn.abisource.com/svnroot/abiword/trunk@30565 bcba8976-2d24-0410-9c9c-aab3bd5fdfd6

+++ src/af/ev/cocoa/ev_CocoaToolbar.cpp
@@ -935,7 +935,7 @@ bool EV_CocoaToolbar::refreshToolbar(AV_View * pView, AV_ChangeMask mask)
 						}
 					}
 					if (value) {
-						int idx = [item indexOfItemWithObjectValue:value];
+						NSInteger idx = [item indexOfItemWithObjectValue:value];
 						if (idx == NSNotFound) {
 							[item setStringValue:value];
 						}
